Use Mixin for StoreOptions and InteropOptions so they don't have an empty heading appear.
Merge DataStorageOptions into BeaconNodeDataOptions.
