+++ teku/src/main/java/tech/pegasys/teku/cli/subcommand/admin/WeakSubjectivityCommand.java
@@ -19,7 +19,6 @@ import org.hyperledger.besu.metrics.noop.NoOpMetricsSystem;
 import picocli.CommandLine;
 import tech.pegasys.teku.cli.converter.PicoCliVersionProvider;
 import tech.pegasys.teku.cli.options.BeaconNodeDataOptions;
-import tech.pegasys.teku.cli.options.DataStorageOptions;
 import tech.pegasys.teku.cli.options.Eth2NetworkOptions;
 import tech.pegasys.teku.service.serviceutils.layout.DataDirLayout;
 import tech.pegasys.teku.spec.Spec;
@@ -60,11 +59,10 @@ public class WeakSubjectivityCommand implements Runnable {
       footerHeading = "%n",
       footer = "Teku is licensed under the Apache License 2.0")
   public int clearWeakSubjectivityState(
-      @CommandLine.Mixin final BeaconNodeDataOptions dataOptions,
-      @CommandLine.Mixin final DataStorageOptions dataStorageOptions,
+      @CommandLine.Mixin final BeaconNodeDataOptions beaconNodeDataOptions,
       @CommandLine.Mixin final Eth2NetworkOptions eth2NetworkOptions)
       throws Exception {
-    try (final Database db = createDatabase(dataOptions, dataStorageOptions, eth2NetworkOptions)) {
+    try (final Database db = createDatabase(beaconNodeDataOptions, eth2NetworkOptions)) {
       // Pull value before updating
       final WeakSubjectivityState original = db.getWeakSubjectivityState();
       if (original.isEmpty()) {
@@ -91,11 +89,10 @@ public class WeakSubjectivityCommand implements Runnable {
       footerHeading = "%n",
       footer = "Teku is licensed under the Apache License 2.0")
   public int displayWeakSubjectivityState(
-      @CommandLine.Mixin final BeaconNodeDataOptions dataOptions,
-      @CommandLine.Mixin final DataStorageOptions dataStorageOptions,
+      @CommandLine.Mixin final BeaconNodeDataOptions beaconNodeDataOptions,
       @CommandLine.Mixin final Eth2NetworkOptions eth2NetworkOptions)
       throws Exception {
-    try (final Database db = createDatabase(dataOptions, dataStorageOptions, eth2NetworkOptions)) {
+    try (final Database db = createDatabase(beaconNodeDataOptions, eth2NetworkOptions)) {
       final WeakSubjectivityState wsState = db.getWeakSubjectivityState();
       SUB_COMMAND_LOG.display("Stored weak subjectivity state: " + stateToString(wsState));
       return 0;
@@ -103,17 +100,17 @@ public class WeakSubjectivityCommand implements Runnable {
   }
 
   private Database createDatabase(
-      final BeaconNodeDataOptions dataOptions,
-      final DataStorageOptions dataStorageOptions,
+      final BeaconNodeDataOptions beaconNodeDataOptions,
       final Eth2NetworkOptions eth2NetworkOptions) {
     final Spec spec = eth2NetworkOptions.getNetworkConfiguration().getSpec();
     final VersionedDatabaseFactory databaseFactory =
         new VersionedDatabaseFactory(
             new NoOpMetricsSystem(),
-            DataDirLayout.createFrom(dataOptions.getDataConfig()).getBeaconDataDirectory(),
-            dataStorageOptions.getDataStorageMode(),
+            DataDirLayout.createFrom(beaconNodeDataOptions.getDataConfig())
+                .getBeaconDataDirectory(),
+            beaconNodeDataOptions.getDataStorageMode(),
             eth2NetworkOptions.getNetworkConfiguration().getEth1DepositContractAddress(),
-            dataStorageOptions.isStoreNonCanonicalBlocks(),
+            beaconNodeDataOptions.isStoreNonCanonicalBlocks(),
             eth2NetworkOptions.getNetworkConfiguration().isEquivocatingIndicesEnabled(),
             spec);
     return databaseFactory.createDatabase();
